+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Rare invalid player movement disconnect
+
 ## [1.2.6] - 2025-08-21
 ### Fixed
 - Potential crash during vanilla texture retrieval